Beam: Cli runs a simple interactive cli, exposed as a beam stream.
Docker-DCO-1.1-Signed-off-by: Solomon Hykes <solomon@docker.com> (github: shykes)
